split scripts
